third_party/blink/web_tests/fast/scrolling/events/scrollend-event-fired-after-snap.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<style>
div {
  position: absolute;
}
#scroller {
  width: 500px;
  height: 500px;
  overflow: scroll;
  scroll-snap-type: both mandatory;
  border: solid black 5px;
}
#space {
  width: 2000px;
  height: 2000px;
}
.target {
  width: 200px;
  height: 200px;
  scroll-snap-align: start;
  background-color: blue;
}
</style>

<body style="margin:0" onload=runTests()>
  <div id="scroller">
    <div id="space"></div>
    <div class="target" style="left: 0px; top: 0px;"></div>
    <div class="target" style="left: 80px; top: 80px;"></div>
    <div class="target" style="left: 200px; top: 200px;"></div>
  </div>
</body>

<script>
if (window.internals)
  internals.runtimeFlags.overscrollCustomizationEnabled = true;

const scroller = document.getElementById("scroller");
const space = document.getElementById("space");
const pixels_to_scroll = 200;

// Drag upwards to scroll down.
const drag = {
  start_x: 50,
  start_y: 230,
  end_x: 50,
  end_y: 50
};

function scrollEndCondition() {
  return !scroll_arrived_after_scroll_end;
}

let scroll_arrived_after_scroll_end;
let scroll_end_arrived;
let scroll_end_promise;

async function resetTest(t) {
  scroller.scrollTo(0, 0);
  await waitForCompositorCommit();

  // Wait 2 additional rAFs before resetting the flags to avoid resetting
  // before the scrollend associated with the scrollTo call.
  await raf();
  await raf();

  scroll_end_arrived = false;
  scroll_arrived_after_scroll_end = false;

  scroll_end_promise = waitForScrollendEvent(scroller).then(() => {
    scroll_end_arrived = true;
  });

  const scrollListener = () => {
    if (scroll_end_arrived)
      scroll_arrived_after_scroll_end = true;
  }
  scroller.addEventListener('scroll', scrollListener);
  t.add_cleanup(() =>  {
    scroller.removeEventListener('scroll', scrollListener);
  });
}

window.onload = () => {
  promise_test (async (t) => {
    resetTest(t);

    await touchDragTo(drag);
    await scroll_end_promise;

    // Verify that scroll snap animation has finished before firing scrollend
    // event.
    assert_equals(scroller.scrollTop, pixels_to_scroll);
    return conditionHolds(scrollEndCondition);
  }, "Tests that scrollend is fired after scroll snap animation completion.");

  promise_test (async (t) => {
    resetTest(t);

    await touchFling(drag);
    await scroll_end_promise;

    // Verify that scroll snap animation has finished before firing scrollend
    // event.
    assert_equals(scroller.scrollTop, pixels_to_scroll);
    return conditionHolds(scrollEndCondition);
  }, "Tests that scrollend is fired after fling snap animation completion.");
};
</script>
